drm: amdgpu: psp: enhance AMDGPU driver to support PSP component.
Enhance AMDGPU driver to support PSP component.

PSP (Platform Security Processor) is one of the IP block on
the Graphics North Bridge.

PSP runs a secure OS and trusted applications in a secure world.

It provides mailbox registers over the Graphics BAR, for
communication with the Rich world.

The mailbox registers are accessible over the PCI interface.

PSP component driver (G2P - Graphics to PSP) enables the
host communication with the PSP.

Signed-off-by: Maulik Mankad <maulik.mankad@amd.com>
Signed-off-by: Sanjay R Mehta <sanju.mehta@amd.com>