drm: hdcpss - add hdcp security service component
HDCP securty service component plays an important role
during the authentication phase of the HDCP 1.4/2.2
receiver with the HDCP transmitter. The HDCP transmitter
and the receiver exchange a series of security messages
to establish the authenticity of the receiver. The
receiver that is successfully authenticated by the
transmitter is authorized to play the protected content
that is sent by the transmitter.

The HDCP Trusted Application(TA) running in PSP secure
world provides for methods to authenticate and authorize
the HDCP receiver.

The HDCP TA communicates with the HDCP receiver through
HDCP SS component, that acts as a bridge between the two
communicating agents.

HDCPSS component uses DAL's I2C communication service for
programming HDMI registers.

HDCPSS component uses the TCI interface to communicate
with the HDCP Trustlet application running on PSP.

Signed-off-by: Maulik Mankad <maulik.mankad@amd.com>
Signed-off-by: Sagar Alatgi <Sagar-mahesh.Alatgi@amd.com>
Signed-off-by: Sanjay R Mehta <sanju.mehta@amd.com>