drm: hdcpss - authenticate only HDCP compliant displays
When HDCP security service receives a hotplug notification,
it triggers the HDCP authentication sequence with the
connected display.

This sequence should be triggered only for HDCP compliant
displays.

A display is HDCP compliant if BKSV consists of 20 1's.

Based on this, the HDCPSS triggers authentication only
if it finds that the connected display is HDCP compliant.

Signed-off-by: Maulik Mankad <maulik.mankad@amd.com>
Signed-off-by: Sagar Alatgi <Sagar-mahesh.Alatgi@amd.com>

Change-Id: I49dc3ff7081964c2a50973c3985f98b22edcff78
Signed-off-by: Sanjay R Mehta <sanju.mehta@amd.com>