drm: hdcpss - update content protection status property
Content protection status property provides a way to userspace
to know the current protection status i.e. whether encryption
is on or off.

When authentication is completed successfully and encryption
is turned ON, this property is updated with a value as 1.

If encryption is OFF, the value read will be 0.

Change-Id: I4c9709b6128f37951a0ab0a0440938afbafc28ee
Signed-off-by: Maulik Mankad <maulik.mankad@amd.com>
Signed-off-by: Sanjay R Mehta <sanju.mehta@amd.com>